[Feat] Add roles logic

+++ src/main/java/com/lovejazz/gymsession/repository/TrainingDiaryRepository.java
@@ -20,7 +20,8 @@ public class TrainingDiaryRepository {
 
     public List<TrainingDiaryDTO> findAll() {
         return jdbcClient.sql("SELECT * FROM Training_Diary")
-                .query((rs, rowNum) -> {;
+                .query((rs, rowNum) -> {
+                    ;
                     return new TrainingDiaryDTO(rs.getInt("id"),
                             rs.getInt("user_id"),
                             rs.getString("name"),
@@ -29,9 +30,10 @@ public class TrainingDiaryRepository {
                 .list();
     }
 
-    public Optional<TrainingDiaryDTO> findById(@RequestParam Integer id) {
-        return jdbcClient.sql("SELECT * FROM Training_Diary WHERE id = :id")
+    public Optional<TrainingDiaryDTO> findById(@RequestParam Integer id, Integer userId) {
+        return jdbcClient.sql("SELECT * FROM Training_Diary WHERE id = :id AND user_id = :userId")
                 .param("id", id)
+                .param("userId", userId)
                 .query((rs, rowNum) -> new TrainingDiaryDTO(rs.getInt("id"),
                         rs.getInt("user_id"),
                         rs.getString("name"),
@@ -40,23 +42,23 @@ public class TrainingDiaryRepository {
     }
 
 
-    public void create(TrainingDiaryDTO diary) {
+    public void create(TrainingDiaryDTO diary, Integer userId) {
         var created = jdbcClient.sql("INSERT INTO Training_Diary(id,user_id,name,sport_type_id) values(?,?,?,?)")
-                .params(List.of(diary.id(), 1, diary.name(), diary.sportTypeId()))
+                .params(List.of(diary.id(), userId, diary.name(), diary.sportTypeId()))
                 .update();
         Assert.state(created == 1, "Failed to create run " + diary.name());
 
     }
 
-    public void update(TrainingDiaryDTO diary, Integer id) {
+    public void update(TrainingDiaryDTO diary, Integer diaryId, Integer userId) {
         var updated = jdbcClient.sql("UPDATE Training_Diary set id = ?, user_id = ?, name = ?, sport_type_id = ? where id = ?")
-                .params(List.of(diary.id(), 1, diary.name(), diary.sportTypeId(), id))
+                .params(List.of(diary.id(), userId, diary.name(), diary.sportTypeId(), diaryId))
                 .update();
         Assert.state(updated == 1, "Failed to update run " + diary.name());
     }
 
-    public void delete(Integer id) {
-        var deleted = jdbcClient.sql("DELETE from Training_Diary WHERE id = :id").param("id", id).update();
+    public void delete(Integer id, Integer userId) {
+        var deleted = jdbcClient.sql("DELETE from Training_Diary WHERE id = :id and user_id = :userId").param("id", id).param("user_id", userId).update();
     }
 
 }
